Add files via upload
Flip-Flop Circuit
FLIP-FLOP

There are electronic elements whose outputs are stable in two ways with the signal applied to the flip flop circuit inputs. Multivibrators that can maintain the state and, accordingly, the value at one of its outputs continuously when an operating voltage is applied to the circuit are called flip flop circuits. They are elements with different output values ​​in response to variable input voltages, which also contain logic gates in their structure.

 

 The instantaneous change of flip flop circuits is called triggering. It has a total of two stable states, SET (1) and Reset (0). These circuit elements remain in one of the two steady states. For this reason, flip flop circuits are used as storage elements. They maintain their output status as long as the signal applied to their input is not changed, and they can also store 1-bit information.

 

How Does a Flip Flop Circuit Work?

 

 The circuit is formed by the capacitors filling up and giving voltage in the opposite direction and cutting the connection of these leds with GND by giving reverse voltage to the base of the NPN transistors, which are normally working at saturation, that is, active. The flashing times of the LEDs in this circuit are related to the capacitors. If you then use higher Farad capacitors, generally the flashing time will be longer. The LEDs do not light when there is no voltage to turn the transistors on. In addition, capacitors are normally empty.